+++ src/common/interfaces/table.ts
@@ -1,0 +1,7 @@
+import { TableRow } from './table-row';
+
+export interface Table {
+  firstCellIsHeader?: boolean;
+  head: TableRow;
+  rows: TableRow[];
+}