+++ src/professions/admin/add-profession/confirmation.controller.ts
@@ -30,7 +30,7 @@ export class ConfirmationController {
 
   @Get('/:id/confirmation')
   @Permissions(UserPermission.CreateProfession)
-  @Render('professions/admin/add-profession/confirmation')
+  @Render('admin/professions/add-profession/confirmation')
   async new(@Param('id') id: string): Promise<ConfirmationTemplate> {
     const profession = await this.professionsService.find(id);