+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Add more granular database fields for storing the duration of a qualification
 - Add a service owner boolean to a user
 - Define new user permissions
+- List Regulatory Authorities
 
 ### Changed